# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 2.0.0
  description: PharmaPOS Pakistan - Complete Pharmacy Management System API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /suppliers:
    get:
      operationId: getSuppliers
      tags: [suppliers]
      responses:
        "200":
          description: Suppliers
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Supplier"
    post:
      operationId: createSupplier
      tags: [suppliers]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateSupplierInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Supplier"
  /suppliers/{id}:
    put:
      operationId: updateSupplier
      tags: [suppliers]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateSupplierInput"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Supplier"
    delete:
      operationId: deleteSupplier
      tags: [suppliers]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
...
  /sales:
    get:
      operationId: getSales
      tags: [sales]
      parameters:
        - name: startDate
          in: query
          schema:
            type: string
        - name: endDate
          in: query
          schema:
            type: string
        - name: customerId
          in: query
          schema:
            type: integer
        - name: paymentStatus
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Sales
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Sale"
    post:
      operationId: createSale
      tags: [sales]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateSaleInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Sale"
  /sales/{id}:
    get:
      operationId: getSale
      tags: [sales]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Sale
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Sale"
...
  /reports/sales-summary:
    get:
      operationId: getSalesSummary
      tags: [reports]
      parameters:
        - name: period
          in: query
          schema:
            type: string
            enum: [today, week, month, year]
      responses:
        "200":
          description: Sales summary
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SalesSummary"
  /reports/top-medicines:
    get:
      operationId: getTopMedicines
      tags: [reports]
      parameters:
        - name: period
          in: query
          schema:
            type: string
            enum: [today, week, month, year]
      responses:
        "200":
          description: Top medicines
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/TopMedicine"
  /reports/expiring-medicines:
    get:
      operationId: getExpiringMedicines
      tags: [reports]
      parameters:
        - name: days
          in: query
          schema:
            type: integer
            default: 30
      responses:
        "200":
          description: Expiring medicines
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Medicine"
...
components:
  schemas:
...
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
      required: [success, message]
...
    Medicine:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        genericName:
          type: string
        categoryId:
          type: integer
        categoryName:
          type: string
        supplierId:
          type: integer
        supplierName:
          type: string
        batchNumber:
          type: string
        barcode:
          type: string
        purchasePrice:
          type: number
        salePrice:
          type: number
        stockQuantity:
          type: integer
        minStockLevel:
          type: integer
        unit:
          type: string
        expiryDate:
          type: string
        manufacturingDate:
          type: string
        manufacturer:
          type: string
        description:
          type: string
        requiresPrescription:
          type: boolean
        isActive:
          type: boolean
        createdAt:
          type: string
      required: [id, name, salePrice, stockQuantity, minStockLevel, unit, isActive, createdAt]
...
    Supplier:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        contactPerson:
          type: string
        phone:
          type: string
        email:
          type: string
        address:
          type: string
        city:
          type: string
        ntn:
          type: string
        createdAt:
          type: string
      required: [id, name, phone, createdAt]
    CreateSupplierInput:
      type: object
      properties:
        name:
          type: string
        contactPerson:
          type: string
        phone:
          type: string
        email:
          type: string
        address:
          type: string
        city:
          type: string
        ntn:
          type: string
      required: [name, phone]
...
    SaleItem:
      type: object
      properties:
        id:
          type: integer
        medicineId:
          type: integer
        medicineName:
          type: string
        medicineUnit:
          type: string
        quantity:
          type: integer
        unitPrice:
          type: number
        discount:
          type: number
        total:
          type: number
      required: [id, medicineId, medicineName, medicineUnit, quantity, unitPrice, discount, total]
    Sale:
      type: object
      properties:
        id:
          type: integer
        invoiceNumber:
          type: string
        customerId:
          type: integer
        customerName:
          type: string
        items:
          type: array
          items:
            $ref: "#/components/schemas/SaleItem"
        subtotal:
          type: number
        discount:
          type: number
        tax:
          type: number
        totalAmount:
          type: number
        paidAmount:
          type: number
        changeAmount:
          type: number
        creditAmount:
          type: number
        paymentStatus:
          type: string
          enum: [paid, credit, partial]
        paymentMethod:
          type: string
          enum: [cash, card, jazzcash, easypaisa, credit]
        status:
          type: string
        notes:
          type: string
        createdAt:
          type: string
      required: [id, invoiceNumber, items, subtotal, discount, tax, totalAmount, paidAmount, changeAmount, creditAmount, paymentStatus, paymentMethod, status, createdAt]
    CreateSaleInput:
      type: object
      properties:
        customerId:
          type: integer
        items:
          type: array
          items:
            type: object
            properties:
              medicineId:
                type: integer
              quantity:
                type: integer
              discount:
                type: number
            required: [medicineId, quantity]
        discount:
          type: number
        paidAmount:
          type: number
        paymentMethod:
          type: string
          enum: [cash, card, jazzcash, easypaisa, credit]
        notes:
          type: string
      required: [items, paidAmount, paymentMethod]
...
    SalesSummary:
      type: object
      properties:
        totalRevenue:
          type: number
        totalTransactions:
          type: integer
        averageTransaction:
          type: number
        dailyBreakdown:
          type: array
          items:
            type: object
            properties:
              date:
                type: string
              revenue:
                type: number
              transactions:
                type: integer
            required: [date, revenue, transactions]
      required: [totalRevenue, totalTransactions, averageTransaction, dailyBreakdown]
    TopMedicine:
      type: object
      properties:
        medicineId:
          type: integer
        medicineName:
          type: string
        quantitySold:
          type: integer
        revenue:
          type: number
      required: [medicineId, medicineName, quantitySold, revenue]
...